Using an int instead of a byte

+++ plugin/src/main/java/net/codingarea/challenges/plugin/management/scheduler/Scheduled.java
@@ -31,7 +31,7 @@ public @interface Scheduled {
 	}
 
 	@Nonnegative
-	byte ticks();
+	int ticks();
 
 	boolean async() default true;
